Refactor project structure and format code with linter
