+++ assignments/assignment_2/src/backend/profile_service/DBConn.js
@@ -14,7 +14,7 @@ module.exports = function () {
     db.connect(function (err) {
         if (err) {
             console.log("DB Request Failed")
-        } 
+        }
     });
 
     return db
