+++ assignments/assignment_2/src/frontend/components/auth_components/ForgotPassword.js
@@ -1,6 +1,6 @@
 import React, { useRef, useState } from 'react';
 import { Form, Button, Card, Alert } from 'react-bootstrap';
-import { useAuth } from '../../contexts/AuthContext';
+import { useAuth } from '../../../contexts/AuthContext';
 import { Link } from 'react-router-dom';
 
 export default function ForgotPassword() {
